refactor(core): reafctoring app
