+++ src/app/core/port/user.service.ts
@@ -1,12 +1,13 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { UserFirebaseService } from './user-firebase.service';
 import { User } from '../entity/user.interface';
+import { UserFirebaseService } from '../adapter/user-firebase.service';
 
 @Injectable({
   providedIn: 'root',
   useClass: UserFirebaseService,
 })
 export abstract class UserService {
+   abstract fetch(userId: string, bearerToken: string): Observable<User>;
   abstract create(user: User, bearerToken: string): Observable<void>;
 }
