+++ src/app/visitor/signup/domain/email-already-taken-error.spec.ts
@@ -3,6 +3,7 @@ import { TestBed } from '@angular/core/testing';
 import { Router } from '@angular/router';
 import { UserService } from '@app/core/port/user.service';
 import { UserStore } from '@app/core/store/user.store';
+import { AuthenticationService } from '@app/core/port/authentication.service';
 import { RegisterUserUseCase } from './register-user.use-case';
 
 describe('RegisterUserUseCaseService', () => {
