+++ src/app/visitor/signup/signup.page.component.ts
@@ -1,18 +1,18 @@
 import { CommonModule } from '@angular/common';
 import {
-    ChangeDetectionStrategy,
-    Component,
-    computed,
-    inject,
-    signal,
+  ChangeDetectionStrategy,
+  Component,
+  computed,
+  inject,
+  signal,
 } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { Router } from '@angular/router';
 import { Visitor } from '@app/core/entity/user.interface';
 import { AuthenticationService } from '@app/core/port/authentication.service';
 import { UserStore } from '@app/core/store/user.store';
+import { RegisterUserUseCaseService } from './domain/register-user.use-case.service';
 import { EmailAlreadyTakenError } from './email-already-taken.error';
-import { RegisterUserUseCaseService } from './register-user.use-case.service';
 
 @Component({
   selector: 'app-signup',
